mcp-toolbox/mcp-toolbox/cloud_functions/ocr_extraction/requirements.txt
functions-framework
google-cloud-storage
pytesseract
python-magic
tika